+++ packages/react-components/react-dialog/src/components/DialogSurface/renderDialogSurface.tsx
@@ -1,10 +1,12 @@
 /** @jsxRuntime automatic */
 /** @jsxImportSource @fluentui/react-jsx-runtime */
 
-import { assertSlots } from '@fluentui/react-utilities';
-import type { DialogSurfaceState, DialogSurfaceSlots, DialogSurfaceContextValues } from './DialogSurface.types';
-import { DialogSurfaceProvider } from '../../contexts';
 import { Portal } from '@fluentui/react-portal';
+import { assertSlots } from '@fluentui/react-utilities';
+
+import { DialogSurfaceProvider } from '../../contexts';
+import { DialogBackdropMotion } from '../DialogBackdropMotion';
+import type { DialogSurfaceState, DialogSurfaceSlots, DialogSurfaceContextValues } from './DialogSurface.types';
 
 /**
  * Render the final JSX of DialogSurface
@@ -14,7 +16,11 @@ export const renderDialogSurface_unstable = (state: DialogSurfaceState, contextV
 
   return (
     <Portal noVirtualParent={state.noVirtualParent} mountNode={state.mountNode}>
-      {state.backdrop && <state.backdrop />}
+      {state.backdrop && (
+        <DialogBackdropMotion appear visible={state.open}>
+          <state.backdrop />
+        </DialogBackdropMotion>
+      )}
       <DialogSurfaceProvider value={contextValues.dialogSurface}>
         <state.root />
       </DialogSurfaceProvider>