+++ packages/react-components/react-motions-preview/stories/CreateMotionComponent/CreateMotionComponentDefault.stories.tsx
@@ -1,4 +1,4 @@
-import { makeStyles, shorthands, tokens } from '@fluentui/react-components';
+import { makeStyles, tokens } from '@fluentui/react-components';
 import { createMotionComponent, motionTokens } from '@fluentui/react-motions-preview';
 import type { MotionImperativeRef } from '@fluentui/react-motions-preview';
 import * as React from 'react';
@@ -8,16 +8,15 @@ const useClasses = makeStyles({
     display: 'flex',
   },
   card: {
-    ...shorthands.border(tokens.strokeWidthThicker, 'solid', tokens.colorNeutralForeground3),
-    ...shorthands.borderRadius(tokens.borderRadiusMedium),
-
-    ...shorthands.padding('20px'),
-    ...shorthands.margin('20px'),
+    border: `${tokens.strokeWidthThicker} solid ${tokens.colorNeutralForeground3}`,
+    borderRadius: tokens.borderRadiusMedium,
+    padding: '20px',
+    margin: '20px',
   },
   item: {
     backgroundColor: tokens.colorBrandBackground,
-    ...shorthands.border(tokens.strokeWidthThicker, 'solid', tokens.colorTransparentStroke),
-    ...shorthands.borderRadius('50%'),
+    border: `${tokens.strokeWidthThicker} solid ${tokens.colorTransparentStroke}`,
+    borderRadius: '50%',
 
     width: '100px',
     height: '100px',