+++ packages/react-components/react-nav-preview/src/components/NavItem/NavItem.types.ts
@@ -1,8 +1,9 @@
-import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
 import { NavItemValue } from '../NavContext.types';
+import type { ARIAButtonSlotProps } from '@fluentui/react-aria';
+import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
 
 export type NavItemSlots = {
-  root: NonNullable<Slot<'a'>>;
+  root: NonNullable<Slot<ARIAButtonSlotProps<'a'>>>;
 
   /**
    * Icon that renders before the content.
@@ -13,7 +14,8 @@ export type NavItemSlots = {
 /**
  * NavItem Props
  */
-export type NavItemProps = ComponentProps<Partial<NavItemSlots>> & {
+export type NavItemProps = ComponentProps<NavItemSlots> & {
+  href?: string;
   /**
    * The value that identifies this navCategoryItem when selected.
    */