+++ packages/react-components/react-portal-compat/CHANGELOG.md
@@ -1,8 +1,37 @@
 # Change Log - @fluentui/react-portal-compat
 
-This log was last generated on Mon, 13 May 2024 12:34:20 GMT and should not be manually modified.
+This log was last generated on Thu, 23 May 2024 08:02:52 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.0.147](https://github.com/microsoft/fluentui/tree/@fluentui/react-portal-compat_v9.0.147)
+
+Thu, 23 May 2024 08:02:52 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-portal-compat_v9.0.146..@fluentui/react-portal-compat_v9.0.147)
+
+### Patches
+
+- Bump @fluentui/react-tabster to v9.21.4 ([commit](https://github.com/microsoft/fluentui/commit/03599d609e8310b08c57d1f871cffbf717d79207) by beachball)
+- Bump @fluentui/react-components to v9.53.0 ([commit](https://github.com/microsoft/fluentui/commit/03599d609e8310b08c57d1f871cffbf717d79207) by beachball)
+
+## [9.0.146](https://github.com/microsoft/fluentui/tree/@fluentui/react-portal-compat_v9.0.146)
+
+Mon, 20 May 2024 12:45:09 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-portal-compat_v9.0.145..@fluentui/react-portal-compat_v9.0.146)
+
+### Patches
+
+- Bump @fluentui/react-tabster to v9.21.3 ([PR #26682](https://github.com/microsoft/fluentui/pull/26682) by beachball)
+- Bump @fluentui/react-components to v9.52.0 ([PR #26682](https://github.com/microsoft/fluentui/pull/26682) by beachball)
+
+## [9.0.145](https://github.com/microsoft/fluentui/tree/@fluentui/react-portal-compat_v9.0.145)
+
+Thu, 16 May 2024 09:25:20 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-portal-compat_v9.0.144..@fluentui/react-portal-compat_v9.0.145)
+
+### Patches
+
+- Bump @fluentui/react-components to v9.51.1 ([PR #31390](https://github.com/microsoft/fluentui/pull/31390) by beachball)
 
 ## [9.0.144](https://github.com/microsoft/fluentui/tree/@fluentui/react-portal-compat_v9.0.144)